+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/network/RetrofitFactory.kt
@@ -18,6 +18,7 @@ package im.vector.matrix.android.internal.network
 
 import com.squareup.moshi.Moshi
 import dagger.Lazy
+import im.vector.matrix.android.internal.util.ensureTrailingSlash
 import okhttp3.Call
 import okhttp3.OkHttpClient
 import okhttp3.Request
@@ -29,7 +30,7 @@ class RetrofitFactory @Inject constructor(private val moshi: Moshi) {
 
     fun create(okHttpClient: Lazy<OkHttpClient>, baseUrl: String): Retrofit {
         return Retrofit.Builder()
-                .baseUrl(baseUrl)
+                .baseUrl(baseUrl.ensureTrailingSlash())
                 .callFactory(object : Call.Factory {
                     override fun newCall(request: Request): Call {
                         return okHttpClient.get().newCall(request)